+++ src/onyx/log/commands/seal_output.clj
@@ -45,4 +45,4 @@
 
 (s/defmethod extensions/fire-side-effects! :seal-output :- State
   [{:keys [args]} old new diff state]
-  (common/start-new-lifecycle old new diff state))
+  (common/start-new-lifecycle old new diff state :job-completed))