+++ changes.md
@@ -4,6 +4,9 @@
 - **API breaking change**: `:aggregation/fn` was renamed to `:aggregation/create-state-update`
 - **API breaking change**: changed the signatures of the following aggregation function types: `default-window-state`, `apply-state-update`, `super-aggregation-fn`. All now take the window as the first argument.
 - **API breaking change**: internal messaging functions no longer take the event map as an argument. Note, this will break any plugins that manually manage the ack count, such as plugins using async callbacks.
+- **API breaking change**: triggers / refinements are now resolved to vars via namespace lookup e.g.
+  Trigger example: `:trigger/on :segment` -> `:trigger/on :onyx.triggers/segment`,
+  Refinement example: `:trigger/refinement :accumulating` -> `:trigger/refinement :onyx.refinements/accumulating`,
 - Bug fix: fixed exceptions losing their main `.getCause` message when thrown inside a lifecycle
 - New feature: Jobs now support metadata by including a metadata map with job submission data e.g. `{:workflow ... :catalog ... :job-metadata {:name "myjobname"}}`. This makes it easier to correlate information about jobs, query the cluster for jobs, etc.
 - Design change: Implemented trigger refinements in terms of state updates. In order to implement a refinement, one must now implement a refinement calls map, analagous to the aggregation calls map. This takes the form `{:refinement/create-state-update (fn [trigger state state-event]) :refinement/apply-state-update (fn [trigger state entry])}`